+++ src/commands/migrator.ts
@@ -11,9 +11,11 @@ import {
   COPILOT_COMMAND_LAST_USED,
   COPILOT_COMMAND_MODEL_KEY,
   COPILOT_COMMAND_SLASH_ENABLED,
+  DEFAULT_COMMANDS,
 } from "@/commands/constants";
 import { COPILOT_COMMAND_CONTEXT_MENU_ENABLED } from "@/commands/constants";
 import { ConfirmModal } from "@/components/modals/ConfirmModal";
+import { customCommandsAtom, customCommandsStore } from "@/commands/state";
 
 function saveUnsupportedCommands(commands: CustomCommand[]) {
   const folderPath = getCustomCommandsFolder();
@@ -83,4 +85,34 @@ export async function migrateCommands() {
   updateSetting("inlineEditCommands", []);
 
   new ConfirmModal(app, () => {}, message, "Commands migrated successfully", "OK", "").open();
+}
+
+export async function generateDefaultCommands(): Promise<void> {
+  const existingCommands = customCommandsStore.get(customCommandsAtom);
+  const defaultCommands = DEFAULT_COMMANDS.filter(
+    (command) => !existingCommands.some((c) => c.title === command.title)
+  );
+  const newCommands = [...existingCommands, ...defaultCommands];
+  CustomCommandManager.getInstance().updateCommands(newCommands);
+}
+
+export async function suggestDefaultCommands(): Promise<void> {
+  const suggestedCommand = getSettings().suggestedDefaultCommands;
+  if (suggestedCommand) {
+    return;
+  }
+  const existingCommands = customCommandsStore.get(customCommandsAtom);
+  if (existingCommands.length === 0) {
+    new ConfirmModal(
+      app,
+      () => {
+        generateDefaultCommands();
+      },
+      "Would you like to create some helpful default commands in your custom prompts folder? These commands will be available through the right-click context menu and slash commands in chat.",
+      "Welcome to Copilot",
+      "Create Commands",
+      "Skip"
+    ).open();
+    updateSetting("suggestedDefaultCommands", true);
+  }
 }
